ci: add multi-arch build verification workflow for Loki binary (#20502)

+++ pkg/loghttp/push/push_test.go
@@ -655,7 +655,7 @@ func TestNegativeSizeHandling(t *testing.T) {
 	linesIngested.Reset()
 
 	// Create a custom request parser that will generate negative sizes
-	var mockParser RequestParser = func(_ string, _ *http.Request, _ Limits, _ *runtime.TenantConfigs, _ int, _ int, _ UsageTracker, _ StreamResolver, _ kitlog.Logger) (*logproto.PushRequest, *Stats, error) {
+	var mockParser RequestParser = func(_ string, _ *http.Request, _ Limits, _ *runtime.TenantConfigs, _ int, _ int64, _ UsageTracker, _ StreamResolver, _ kitlog.Logger) (*logproto.PushRequest, *Stats, error) {
 		// Create a minimal valid request
 		req := &logproto.PushRequest{
 			Streams: []logproto.Stream{
@@ -732,7 +732,7 @@ func TestParseRequestWithZeroMaxDecompressedSize(t *testing.T) {
 		contentType          string
 		contentEncoding      string
 		maxRecvMsgSize       int
-		maxDecompressedSize  int
+		maxDecompressedSize  int64
 		expectedError        bool
 		expectedErrorMessage string
 	}{
